Runners/investments-funds-tracker/requirements.txt
google-generativeai
redis[hiredis]
psycopg2-binary
newsapi-python